Re-introducing ability to unset fields.
Previously, by `del model[field_name]` and `model.save()`, a field
could be manually removed from the database. By re-introducing this
behavior / returning to backwards-compatibility, the `Model.save()`
method had to be reverted to a "lossy" replacement versus the
`$set` operation on an `update`. (To be fair, this is more like the
historical `save()` operation anyway.)