Update typing for Python 3.9+ and PyMongo 4.1+
PyMongo 4.1 includes type annotations as part of the release, so there is
no need to maintain a separate (potentially conflicting) set of type stubs.
Additionally, with the release of Python 3.11, we can move support to 3.9+
which means removing many of the older typing workarounds.

This commit includes minimal functional changes to improve typing and bring
the use of PyMongo up to recent versions.